Change default build to O2
This works for me and it makes the UI feel a little snappier (could be placebo, not sure if the UI is firmware or screen hardware limited). 

If I have a large number of local changes or custom images I find the firmware won't fit in memory with O3. I've made it O2 for now to save any future headaches from new features.