Refined UI rendering to avoid redrawing elements unnecessarily.
The UI was excessively redrawn when screens change with the call to
uui_refresh. This causes UI elements to flicker. Instead, each screen
will need to refresh only what is necessary on activation

The settings screen no longer flicker when moving through pages as only
the table portion is blanked with a tft_fill call rather than tft_clear.