Add Watts display and implement original CC/CV behaviour.
This replaces CC / CV / CL with a CCCV screen.
The use of a SW i_limit is replaced with the HW DAC for limiting current.
This makes the unit behave much like the original, without the SW limit cutting power.
The SW limit is now set to the maximum output current supported by the DPS model.