opendps/uframe: check for overflow (#266)
* opendps/uframe: check for overflow

Instead of calling `memcpy` with a too large payload, bail out and
report the error when the received uframe is too large.

* opendps: Extract payload data in-place

The use of frame_t makes little sense for upgrade data frames, as they
are just to large to `memcpy()` easily. In addition, this allows the
use of way larger chunks than `MAX_FRAME_LENGTH` for the data transfer.